+++ CHANGELOG.md
@@ -5,7 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.7.0] - 2026-01-27
+## [0.7.1] - 2026-01-31
+
+### Changes
+- changed eslint config to ignore linting in `node_modules` and `security-patches`
+- the speed of the `HelpwaveLogo` animation can now be controlled
+
+### Security
+- Update and change package versions to remove vulnerabilities
+
+## [0.7.0] - 2026-01-29
 
 ### Added
 - `ColumnSizingWithTargetFeature` and `AutoColumnOrderFeature` to Table for column size fixing and column ordering